+++ CHANGELOG.md
@@ -22,6 +22,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Condensed the `PhmmParamKind` variants in the `visit_params` API (behind `dev-phmm` feature gate)
 - Added row and column accessors to `TransitionParams` (behind `dev-phmm` and `alignment-diagnostics` feature gates)
 - Added `WithErrorContext` and `ResultWithErrorContext` traits using the `ErrorWithContext` struct for updated error-handling.
+- Added methods and implementations to iterate over CIGAR views
 
 ### Changed